+++ packages/renderer/src/components/project-delete-modal.vue
